// packages/sfn-demo-toolkit/patches/v0.4/extensions/branding/types.ts
/**
 * Copyright 2026 Salesforce, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface BrandHeroSlide {
    id: string;
    title: string;
    subtitle?: string;
    imageUrl: string;
    imageAlt?: string;
    ctaText?: string;
    ctaLink?: string;
}

export interface BrandFeaturedCard {
    title: string;
    description: string;
    imageUrl?: string;
    imageAlt?: string;
    ctaText?: string;
    ctaLink?: string;
}

export interface BrandContent {
    id: string;
    displayName: string;
    logo: {
        src: string;
        alt: string;
    };
    hero: BrandHeroSlide[];
    featuredCards: {
        primary: BrandFeaturedCard[];
        textOnly?: BrandFeaturedCard;
    };
}